feat: RestaurantItem 분리 및 구현
